Fix message passing compiler warnings
By adding explicit static_casts between integer values where needed.

GIT_ORIGIN_SPP_REV_ID: 09254a65f2b6f547c7baa692ea7e3744e8cad190

+++ score/message_passing/qnx_dispatch/qnx_dispatch_engine.cpp
@@ -363,7 +363,7 @@ void QnxDispatchEngine::ProcessTimerQueue() noexcept
     }
     const auto distance = std::chrono::duration_cast<std::chrono::nanoseconds>(then - Clock::now()).count() + 1;
     struct _itimer itimer{};
-    itimer.nsec = distance;
+    itimer.nsec = static_cast<std::uint64_t>(distance);
     itimer.interval_nsec = 0;
     score::cpp::ignore = os_resources_.timer->TimerSettime(timer_id_, 0, &itimer, nullptr);
 }
